Add contract util
