+++ src/rpc/blockchain.cpp
@@ -41,8 +41,8 @@
 #include <txdb.h>
 #include <util/convert.h>
 #include <qtum/qtumdelegation.h>
-#include <qtum/qtumtoken.h>
 #include <util/tokenstr.h>
+#include <rpc/contract_util.h>
 
 #include <stdint.h>
 
@@ -300,49 +300,6 @@ UniValue blockToJSON(const CBlock& block, const CBlockIndex* tip, const CBlockIn
 
     return result;
 }
-
-//////////////////////////////////////////////////////////////////////////// // qtum
-UniValue executionResultToJSON(const dev::eth::ExecutionResult& exRes)
-{
-    UniValue result(UniValue::VOBJ);
-    result.pushKV("gasUsed", CAmount(exRes.gasUsed));
-    std::stringstream ss;
-    ss << exRes.excepted;
-    result.pushKV("excepted", ss.str());
-    result.pushKV("newAddress", exRes.newAddress.hex());
-    result.pushKV("output", HexStr(exRes.output));
-    result.pushKV("codeDeposit", static_cast<int32_t>(exRes.codeDeposit));
-    result.pushKV("gasRefunded", CAmount(exRes.gasRefunded));
-    result.pushKV("depositSize", static_cast<int32_t>(exRes.depositSize));
-    result.pushKV("gasForDeposit", CAmount(exRes.gasForDeposit));
-    result.pushKV("exceptedMessage", exceptedMessage(exRes.excepted, exRes.output));
-    return result;
-}
-
-UniValue transactionReceiptToJSON(const QtumTransactionReceipt& txRec)
-{
-    UniValue result(UniValue::VOBJ);
-    result.pushKV("stateRoot", txRec.stateRoot().hex());
-    result.pushKV("utxoRoot", txRec.utxoRoot().hex());
-    result.pushKV("gasUsed", CAmount(txRec.cumulativeGasUsed()));
-    result.pushKV("bloom", txRec.bloom().hex());
-    UniValue logEntries(UniValue::VARR);
-    dev::eth::LogEntries logs = txRec.log();
-    for(dev::eth::LogEntry log : logs){
-        UniValue logEntrie(UniValue::VOBJ);
-        logEntrie.pushKV("address", log.address.hex());
-        UniValue topics(UniValue::VARR);
-        for(dev::h256 l : log.topics){
-            topics.push_back(l.hex());
-        }
-        logEntrie.pushKV("topics", topics);
-        logEntrie.pushKV("data", HexStr(log.data));
-        logEntries.push_back(logEntrie);
-    }
-    result.pushKV("log", logEntries);
-    return result;
-}
-////////////////////////////////////////////////////////////////////////////
 
 static UniValue getestimatedannualroi(const JSONRPCRequest& request)
 {
@@ -1230,65 +1187,6 @@ static UniValue getblock(const JSONRPCRequest& request)
 }
 
 ////////////////////////////////////////////////////////////////////// // qtum
-UniValue CallToContract(const UniValue& params)
-{
-    LOCK(cs_main);
-
-    std::string strAddr = params[0].get_str();
-    std::string data = params[1].get_str();
-
-    if(data.size() % 2 != 0 || !CheckHex(data))
-        throw JSONRPCError(RPC_TYPE_ERROR, "Invalid data (data not hex)");
-
-    dev::Address addrAccount;
-    if(strAddr.size() > 0)
-    {
-        if(strAddr.size() != 40 || !CheckHex(strAddr))
-            throw JSONRPCError(RPC_INVALID_ADDRESS_OR_KEY, "Incorrect address");
-
-        addrAccount = dev::Address(strAddr);
-        if(!globalState->addressInUse(addrAccount))
-            throw JSONRPCError(RPC_INVALID_ADDRESS_OR_KEY, "Address does not exist");
-    }
-
-    dev::Address senderAddress;
-    if(params.size() >= 3){
-        CTxDestination qtumSenderAddress = DecodeDestination(params[2].get_str());
-        if (IsValidDestination(qtumSenderAddress)) {
-            const PKHash *keyid = boost::get<PKHash>(&qtumSenderAddress);
-            senderAddress = dev::Address(HexStr(valtype(keyid->begin(),keyid->end())));
-        }else{
-            senderAddress = dev::Address(params[2].get_str());
-        }
-
-    }
-    uint64_t gasLimit=0;
-    if(params.size() >= 4){
-        gasLimit = params[3].get_int64();
-    }
-
-    CAmount nAmount = 0;
-    if (params.size() >= 5){
-        nAmount = AmountFromValue(params[4]);
-        if (nAmount < 0)
-            throw JSONRPCError(RPC_TYPE_ERROR, "Invalid amount for send");
-    }
-
-
-    std::vector<ResultExecute> execResults = CallContract(addrAccount, ParseHex(data), senderAddress, gasLimit, nAmount);
-
-    if(fRecordLogOpcodes){
-        writeVMlog(execResults);
-    }
-
-    UniValue result(UniValue::VOBJ);
-    result.pushKV("address", strAddr);
-    result.pushKV("executionResult", executionResultToJSON(execResults[0].execRes));
-    result.pushKV("transactionReceipt", transactionReceiptToJSON(execResults[0].txRec));
-
-    return result;
-}
-
 UniValue callcontract(const JSONRPCRequest& request)
 {
             RPCHelpMan{"callcontract",
@@ -1339,307 +1237,6 @@ UniValue callcontract(const JSONRPCRequest& request)
             }.Check(request);
  
     return CallToContract(request.params);
-}
-
-UniValue SearchLogs(const UniValue& _params);
-class CallToken : public QtumTokenExec, public QtumToken
-{
-public:
-    CallToken()
-    {
-        setQtumTokenExec(this);
-    }
-
-    bool execValid(const int& func, const bool& sendTo)
-    {
-        if(func == -1 || sendTo)
-            return false;
-        return true;
-    }
-
-    bool execEventsValid(const int &func, const int64_t &fromBlock)
-    {
-        if(func == -1 || fromBlock < 0)
-            return false;
-        return true;
-    }
-
-    bool exec(const bool& sendTo, const std::map<std::string, std::string>& lstParams, std::string& result, std::string&)
-    {
-        if(sendTo)
-            return false;
-
-        UniValue params(UniValue::VARR);
-
-        // Set address
-        auto it = lstParams.find(paramAddress());
-        if(it != lstParams.end())
-            params.push_back(it->second);
-        else
-            return false;
-
-        // Set data
-        it = lstParams.find(paramDatahex());
-        if(it != lstParams.end())
-            params.push_back(it->second);
-        else
-            return false;
-
-        // Set sender
-        it = lstParams.find(paramSender());
-        if(it != lstParams.end())
-            params.push_back(it->second);
-
-        // Get execution result
-        UniValue response = CallToContract(params);
-        if(!response.isObject() || !response.exists("executionResult"))
-            return false;
-        UniValue executionResult = response["executionResult"];
-
-        // Get output
-        if(!executionResult.isObject() || !executionResult.exists("output"))
-            return false;
-        UniValue output = executionResult["output"];
-        result = output.get_str();
-
-        return true;
-    }
-
-    bool execEvents(const int64_t &fromBlock, const int64_t &toBlock, const std::string &eventName, const std::string &contractAddress, const std::string &senderAddress, std::vector<TokenEvent> &result)
-    {
-        UniValue resultVar;
-        if(!searchTokenTx(fromBlock, toBlock, contractAddress, senderAddress, resultVar))
-            return false;
-
-        const UniValue& list = resultVar.get_array();
-        for(size_t i = 0; i < list.size(); i++)
-        {
-            // Search the log for events
-            const UniValue& eventMap = list[i].get_obj();
-            const UniValue& listLog = eventMap["log"].get_array();
-            for(size_t i = 0; i < listLog.size(); i++)
-            {
-                // Skip the not needed events
-                const UniValue& eventLog = listLog[i].get_obj();
-                const UniValue& topicsList = eventLog["topics"].get_array();
-                if(topicsList.size() < 3) continue;
-                if(topicsList[0].get_str() != eventName) continue;
-
-                // Create new event
-                TokenEvent tokenEvent;
-                tokenEvent.address = eventMap["contractAddress"].get_str();
-                tokenEvent.sender = topicsList[1].get_str().substr(24);
-                ToQtumAddress(tokenEvent.sender, tokenEvent.sender);
-                tokenEvent.receiver = topicsList[2].get_str().substr(24);
-                ToQtumAddress(tokenEvent.receiver, tokenEvent.receiver);
-                tokenEvent.blockHash = uint256S(eventMap["blockHash"].get_str());
-                tokenEvent.blockNumber = eventMap["blockNumber"].get_int64();
-                tokenEvent.transactionHash = uint256S(eventMap["transactionHash"].get_str());
-
-                // Parse data
-                std::string data = eventLog["data"].get_str();
-                tokenEvent.value = ToUint256(data);
-
-                result.push_back(tokenEvent);
-            }
-        }
-
-        return true;
-    }
-
-    bool searchTokenTx(const int64_t &fromBlock, const int64_t &toBlock, const std::string &contractAddress, const std::string &senderAddress, UniValue& resultVar)
-    {
-        UniValue params(UniValue::VARR);
-        params.push_back(fromBlock);
-        params.push_back(toBlock);
-
-        UniValue addresses(UniValue::VARR);
-        addresses.push_back(contractAddress);
-        UniValue addressesObj(UniValue::VOBJ);
-        addressesObj.pushKV("addresses", addresses);
-        params.push_back(addressesObj);
-
-        UniValue topics(UniValue::VARR);
-        // Skip the event type check
-        static std::string nullRecord = uint256().ToString();
-        topics.push_back(nullRecord);
-        // Match the log with sender address
-        topics.push_back(senderAddress);
-        // Match the log with receiver address
-        topics.push_back(senderAddress);
-        UniValue topicsObj(UniValue::VOBJ);
-        topicsObj.pushKV("topics", topics);
-        params.push_back(topicsObj);
-
-        resultVar = SearchLogs(params);
-
-        return true;
-    }
-};
-
-void assignJSON(UniValue& entry, const TransactionReceiptInfo& resExec) {
-    entry.pushKV("blockHash", resExec.blockHash.GetHex());
-    entry.pushKV("blockNumber", uint64_t(resExec.blockNumber));
-    entry.pushKV("transactionHash", resExec.transactionHash.GetHex());
-    entry.pushKV("transactionIndex", uint64_t(resExec.transactionIndex));
-    entry.pushKV("outputIndex", uint64_t(resExec.outputIndex));
-    entry.pushKV("from", resExec.from.hex());
-    entry.pushKV("to", resExec.to.hex());
-    entry.pushKV("cumulativeGasUsed", CAmount(resExec.cumulativeGasUsed));
-    entry.pushKV("gasUsed", CAmount(resExec.gasUsed));
-    entry.pushKV("contractAddress", resExec.contractAddress.hex());
-    std::stringstream ss;
-    ss << resExec.excepted;
-    entry.pushKV("excepted",ss.str());
-    entry.pushKV("exceptedMessage", resExec.exceptedMessage);
-    entry.pushKV("bloom", resExec.bloom.hex());
-    entry.pushKV("stateRoot", resExec.stateRoot.hex());
-    entry.pushKV("utxoRoot", resExec.utxoRoot.hex());
-}
-
-void assignJSON(UniValue& logEntry, const dev::eth::LogEntry& log,
-        bool includeAddress) {
-    if (includeAddress) {
-        logEntry.pushKV("address", log.address.hex());
-    }
-
-    UniValue topics(UniValue::VARR);
-    for (dev::h256 hash : log.topics) {
-        topics.push_back(hash.hex());
-    }
-    logEntry.pushKV("topics", topics);
-    logEntry.pushKV("data", HexStr(log.data));
-}
-
-void transactionReceiptInfoToJSON(const TransactionReceiptInfo& resExec, UniValue& entry) {
-    assignJSON(entry, resExec);
-
-    const auto& logs = resExec.logs;
-    UniValue logEntries(UniValue::VARR);
-    for(const auto&log : logs){
-        UniValue logEntry(UniValue::VOBJ);
-        assignJSON(logEntry, log, true);
-        logEntries.push_back(logEntry);
-    }
-    entry.pushKV("log", logEntries);
-}
-
-size_t parseUInt(const UniValue& val, size_t defaultVal) {
-    if (val.isNull()) {
-        return defaultVal;
-    } else {
-        int n = val.get_int();
-        if (n < 0) {
-            throw JSONRPCError(RPC_INVALID_PARAMS, "Expects unsigned integer");
-        }
-
-        return n;
-    }
-}
-
-int parseBlockHeight(const UniValue& val) {
-    if (val.isStr()) {
-        auto blockKey = val.get_str();
-
-        if (blockKey == "latest") {
-            return latestblock.height;
-        } else {
-            throw JSONRPCError(RPC_INVALID_PARAMS, "invalid block number");
-        }
-    }
-
-    if (val.isNum()) {
-        int blockHeight = val.get_int();
-
-        if (blockHeight < 0) {
-            return latestblock.height;
-        }
-
-        return blockHeight;
-    }
-
-    throw JSONRPCError(RPC_INVALID_PARAMS, "invalid block number");
-}
-
-int parseBlockHeight(const UniValue& val, int defaultVal) {
-    if (val.isNull()) {
-        return defaultVal;
-    } else {
-        return parseBlockHeight(val);
-    }
-}
-
-dev::h160 parseParamH160(const UniValue& val) {
-    if (!val.isStr()) {
-        throw JSONRPCError(RPC_INVALID_PARAMS, "Invalid hex 160");
-    }
-
-    auto addrStr = val.get_str();
-
-    if (addrStr.length() != 40 || !CheckHex(addrStr)) {
-        throw JSONRPCError(RPC_INVALID_PARAMS, "Invalid hex 160 string");
-    }
-    return dev::h160(addrStr);
-}
-
-void parseParam(const UniValue& val, std::vector<dev::h160> &h160s) {
-    if (val.isNull()) {
-        return;
-    }
-
-    // Treat a string as an array of length 1
-    if (val.isStr()) {
-        h160s.push_back(parseParamH160(val.get_str()));
-        return;
-    }
-
-    if (!val.isArray()) {
-        throw JSONRPCError(RPC_INVALID_PARAMS, "Expect an array of hex 160 strings");
-    }
-
-    auto vals = val.getValues();
-    h160s.resize(vals.size());
-
-    std::transform(vals.begin(), vals.end(), h160s.begin(), [](UniValue val) -> dev::h160 {
-        return parseParamH160(val);
-    });
-}
-
-void parseParam(const UniValue& val, std::set<dev::h160> &h160s) {
-    std::vector<dev::h160> v;
-    parseParam(val, v);
-    h160s.insert(v.begin(), v.end());
-}
-
-void parseParam(const UniValue& val, std::vector<boost::optional<dev::h256>> &h256s) {
-    if (val.isNull()) {
-        return;
-    }
-
-    if (!val.isArray()) {
-        throw JSONRPCError(RPC_INVALID_PARAMS, "Expect an array of hex 256 strings");
-    }
-
-    auto vals = val.getValues();
-    h256s.resize(vals.size());
-
-    std::transform(vals.begin(), vals.end(), h256s.begin(), [](UniValue val) -> boost::optional<dev::h256> {
-        if (val.isNull()) {
-            return boost::optional<dev::h256>();
-        }
-
-        if (!val.isStr()) {
-            throw JSONRPCError(RPC_INVALID_PARAMS, "Invalid hex 256 string");
-        }
-
-        auto addrStr = val.get_str();
-
-        if (addrStr.length() != 64 || !CheckHex(addrStr)) {
-            throw JSONRPCError(RPC_INVALID_PARAMS, "Invalid hex 256 string");
-        }
-
-        return boost::optional<dev::h256>(dev::h256(addrStr));
-    });
 }
 
 class WaitForLogsParams {
@@ -1835,125 +1432,6 @@ UniValue waitforlogs(const JSONRPCRequest& request_) {
     result.pushKV("entries", jsonLogs);
     result.pushKV("count", (int) jsonLogs.size());
     result.pushKV("nextblock", curheight + 1);
-
-    return result;
-}
-
-class SearchLogsParams {
-public:
-    size_t fromBlock;
-    size_t toBlock;
-    size_t minconf;
-
-    std::set<dev::h160> addresses;
-    std::vector<boost::optional<dev::h256>> topics;
-
-    SearchLogsParams(const UniValue& params) {
-        std::unique_lock<std::mutex> lock(cs_blockchange);
-
-        setFromBlock(params[0]);
-        setToBlock(params[1]);
-
-        parseParam(params[2]["addresses"], addresses);
-        parseParam(params[3]["topics"], topics);
-
-        minconf = parseUInt(params[4], 0);
-    }
-
-private:
-    void setFromBlock(const UniValue& val) {
-        if (!val.isNull()) {
-            fromBlock = parseBlockHeight(val);
-        } else {
-            fromBlock = latestblock.height;
-        }
-    }
-
-    void setToBlock(const UniValue& val) {
-        if (!val.isNull()) {
-            toBlock = parseBlockHeight(val);
-        } else {
-            toBlock = latestblock.height;
-        }
-    }
-
-};
-
-UniValue SearchLogs(const UniValue& _params)
-{
-    if(!fLogEvents)
-        throw JSONRPCError(RPC_INTERNAL_ERROR, "Events indexing disabled");
-
-    int curheight = 0;
-
-    LOCK(cs_main);
-
-    SearchLogsParams params(_params);
-
-    std::vector<std::vector<uint256>> hashesToBlock;
-
-    curheight = pblocktree->ReadHeightIndex(params.fromBlock, params.toBlock, params.minconf, hashesToBlock, params.addresses);
-
-    if (curheight == -1) {
-        throw JSONRPCError(RPC_INVALID_PARAMETER, "Incorrect params");
-    }
-
-    UniValue result(UniValue::VARR);
-
-    auto topics = params.topics;
-
-    std::set<uint256> dupes;
-
-    for(const auto& hashesTx : hashesToBlock)
-    {
-        for(const auto& e : hashesTx)
-        {
-
-            if(dupes.find(e) != dupes.end()) {
-                continue;
-            }
-            dupes.insert(e);
-
-            std::vector<TransactionReceiptInfo> receipts = pstorageresult->getResult(uintToh256(e));
-
-            for(const auto& receipt : receipts) {
-                if(receipt.logs.empty()) {
-                    continue;
-                }
-
-                if (!topics.empty()) {
-                    for (size_t i = 0; i < topics.size(); i++) {
-                        const auto& tc = topics[i];
-
-                        if (!tc) {
-                            continue;
-                        }
-
-                        for (const auto& log: receipt.logs) {
-                            auto filterTopicContent = tc.get();
-
-                            if (i >= log.topics.size()) {
-                                continue;
-                            }
-
-                            if (filterTopicContent == log.topics[i]) {
-                                goto push;
-                            }
-                        }
-                    }
-
-                    // Skip the log if none of the topics are matched
-                    continue;
-                }
-
-            push:
-
-                UniValue tri(UniValue::VOBJ);
-                transactionReceiptInfoToJSON(receipt, tri);
-                result.push_back(tri);
-            }
-        }
-    }
 
     return result;
 }
